requirements.txt
streamlit
openai-whisper
streamlit-webrtc
pydub
av
numpy